requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Login
psycopg2
Faker
flask-mysqldb
email_validator

# ----------------------
# Production Server
# ----------------------
gunicorn
